This commit includes following: (#742)
1/ Bug fix while populating extensions from a client hello with empty extensions.
2/ Update the return type for some APIs in client hello where we have a possibility of returning -1

+++ tls/s2n_client_hello.h
@@ -41,11 +41,11 @@ int s2n_client_hello_free_parsed_extensions(struct s2n_client_hello *client_hell
 
 extern struct s2n_client_hello *s2n_connection_get_client_hello(struct s2n_connection *conn);
 
-extern uint32_t s2n_client_hello_get_raw_message_length(struct s2n_client_hello *ch);
-extern uint32_t s2n_client_hello_get_raw_message(struct s2n_client_hello *ch, uint8_t *out, uint32_t max_length);
+extern ssize_t s2n_client_hello_get_raw_message_length(struct s2n_client_hello *ch);
+extern ssize_t s2n_client_hello_get_raw_message(struct s2n_client_hello *ch, uint8_t *out, uint32_t max_length);
 
-extern uint32_t s2n_client_hello_get_cipher_suites_length(struct s2n_client_hello *ch);
-extern uint32_t s2n_client_hello_get_cipher_suites(struct s2n_client_hello *ch, uint8_t *out, uint32_t max_length);
+extern ssize_t s2n_client_hello_get_cipher_suites_length(struct s2n_client_hello *ch);
+extern ssize_t s2n_client_hello_get_cipher_suites(struct s2n_client_hello *ch, uint8_t *out, uint32_t max_length);
 
-extern uint32_t s2n_client_hello_get_extensions_length(struct s2n_client_hello *ch);
-extern uint32_t s2n_client_hello_get_extensions(struct s2n_client_hello *ch, uint8_t *out, uint32_t max_length);
+extern ssize_t s2n_client_hello_get_extensions_length(struct s2n_client_hello *ch);
+extern ssize_t s2n_client_hello_get_extensions(struct s2n_client_hello *ch, uint8_t *out, uint32_t max_length);
